+++ packages/salesforcedx-sobjects-faux-generator/src/types/general.ts
@@ -4,6 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import { SObjectShortDescription } from '../describe';
+import { SObject } from './describe';
 
 export enum SObjectCategory {
   ALL = 'ALL',
@@ -15,4 +17,43 @@ export enum SObjectRefreshSource {
   Manual = 'manual',
   Startup = 'startup',
   StartupMin = 'startupmin'
+}
+
+export interface FieldDeclaration {
+  modifier: string;
+  type: string;
+  name: string;
+  comment?: string;
+}
+
+export type SObjectDefinition = Pick<SObject, 'name'> & {
+  fields: FieldDeclaration[];
+};
+
+export interface SObjectDefinitionRetriever {
+  retrieve: (output: SObjectRefreshOutput) => Promise<void>;
+}
+
+export interface SObjectGenerator {
+  generate: (output: SObjectRefreshOutput) => void;
+}
+
+export interface SObjectRefreshOutput {
+  sfdxPath: string;
+  addTypeNames: (names: SObjectShortDescription[]) => void;
+  getTypeNames: () => SObjectShortDescription[];
+  addStandard: (standard: SObjectDefinition[]) => void;
+  getStandard: () => SObjectDefinition[];
+  addCustom: (standard: SObjectDefinition[]) => void;
+  getCustom: () => SObjectDefinition[];
+  setError: (message: string, stack?: string) => void;
+}
+
+export interface SObjectRefreshResult {
+  data: {
+    cancelled: boolean;
+    standardObjects?: number;
+    customObjects?: number;
+  };
+  error?: { message: string; stack?: string };
 }
